max file download

+++ src/main/java/com/example/linker/service/NoteServiceImpl.java
@@ -62,7 +62,7 @@ public class NoteServiceImpl implements NoteService {
 		
 		note.setNumberOfViews(note.getNumberOfViews() + 1);
 
-		if (note.getNumberOfViews() == note.getMaxNumberOfViews()) {
+		if (note.getFile() == null && note.getNumberOfViews() == note.getMaxNumberOfViews()) {
 			noteRepository.delete(note);
 		} else {
 			noteRepository.save(note);
